Size

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long lli;
 	float f;
 
-	printf("size of a char: %ld byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of an int: %ld byte(s)\n", (unsigned long)sizeof(i));
-	printf("size of a long int: %ld byte(s)\n", (unsigned long)sizeof(li));
-	printf("size of a long long int: %ld byte(s)\n", (unsigned long)sizeof(lli));
-	printf("size of a float: %ld byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of a char: %ld byte(s)\n", sizeof(c));
+	printf("size of an int: %ld byte(s)\n", sizeof(i));
+	printf("size of a long int: %ld byte(s)\n", sizeof(li));
+	printf("size of a long long int: %ld byte(s)\n", sizeof(lli));
+	printf("size of a float: %ld byte(s)\n", sizeof(f));
 	return (0);
 }
